OS X 10.4 SDK @ Xcode 3.0
http://adcdownload.apple.com/Developer_Tools/xcode_3.0/xcode_3.0.dmg

+++ MacOSX10.4u.sdk/System/Library/Frameworks/Kernel.framework/Versions/A/Headers/IOKit/graphics/IODisplay.h
@@ -208,7 +208,7 @@ protected:
     UInt32	fCurrentPowerState;
     SInt32	fMinBrightness;
     SInt32	fMaxBrightness;
-    UInt16	fMaxBrightnessLevel[kIODisplayNumPowerStates];
+    UInt8	fMaxBrightnessLevel[kIODisplayNumPowerStates];
 
 public:
     virtual IOService * probe( IOService *, SInt32 * );
